A bunch of dinosaurs were running a relay race. The end goal of this 4x400m relay on a 400m track was to have every dinosaur run around the track and end up where they started. Well, why not have every dinosaur run at the same time then?
In case you didn't realize, this is a metaphor. A read being aligned is a dinosaur running. I am pretty sure I will be the first and last person to ever make that comparison.

Multiple threads being used in background for read alignment, run-time cut dramatically, discovered the secret to the universe, the cake is a lie, 42